tennis/requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
pytz>=2024.1
requests>=2.31.0
